feat(context): :sparkles: add separate folder for context
